SF-128/feat: 파일 업로드/다운로드

+++ src/main/java/com/ssafy/faraway/domain/hotplace/service/HotPlaceService.java
@@ -8,7 +8,7 @@ import com.ssafy.faraway.domain.hotplace.dto.res.HotPlaceListResponseDto;
 import com.ssafy.faraway.domain.hotplace.dto.res.HotPlaceResponseDto;
 
 public interface HotPlaceService {
-    Integer save(HotPlaceSaveRequestDto hotPlaceSaveRequestDto) throws Exception;
+    Long save(HotPlaceSaveRequestDto hotPlaceSaveRequestDto) throws Exception;
     HotPlaceResponseDto findById(Long id) throws Exception;
     Integer updateHit(Long id) throws Exception;
     PagingResponse<HotPlaceListResponseDto> findAllByCondition(SearchCondition searchCondition) throws Exception;
